refactor(cmake): reorganize includes and update paths
* Updated include paths for various components in CMakeLists.txt
* Removed unused includes and organized dependencies
* Added new source files for optimization wrappers
