+++ src/solver/optimisation/include/antares/solver/optimisation/ortools_quadratic_wrapper.h
