+++ src/solver/optimisation/ortools_quadratic_wrapper.cpp
@@ -6,7 +6,7 @@
 #include <vector>
 
 #include <antares/logs/logs.h>
-#include <antares/solver/utils/ortools_quadratic_wrapper.h>
+#include <antares/solver/optimisation/ortools_quadratic_wrapper.h>
 #include <antares/solver/utils/ortools_utils.h>
 
 using Antares::Optimization::SingleOptimOptions;
