+++ src/solver/simulation/include/antares/solver/simulation/sim_structure_probleme_economique.h
@@ -8,7 +8,7 @@
 #include <memory>
 #include <vector>
 
-#include "antares/solver/optimisation/opt_constants.h"
+#include "antares/solver/utils/opt_constants.h"
 #include "antares/solver/optimisation/opt_structure_probleme_a_resoudre.h"
 #include "antares/solver/utils/optimization_statistics.h"
 #include "antares/study/fwd.h"
