+++ src/solver/utils/CMakeLists.txt
@@ -1,8 +1,8 @@
 set(SRC
         include/antares/solver/utils/ortools_utils.h
         ortools_utils.cpp
-        include/antares/solver/utils/ortools_quadratic_wrapper.h
-        ortools_quadratic_wrapper.cpp
+        include/antares/solver/utils/opt_constants.h
+        include/antares/solver/utils/sim_constants.h
         include/antares/solver/utils/filename.h
         filename.cpp
         include/antares/solver/utils/mps_utils.h
@@ -30,8 +30,8 @@ target_link_libraries(utils
         Antares::result_writer
         Antares::optimization-options
         antares-core
-        Antares::antares-solver-simulation
-        model_antares
+        PRIVATE
+        Antares::simu-table-generator
 )
 
 target_include_directories(utils
