+++ src/tests/src/solver/optimisation/CMakeLists.txt
@@ -18,6 +18,7 @@ add_boost_test(unit-tests-for-solver-optimisation
         ${CMAKE_SOURCE_DIR}/src/solver/optimisation/variables/VariableManagerUtils.cpp
         LIBS
         antares-study-system-model
+        Antares::antares-solver-simulation
         Antares::simulation-table
         Antares::linear-problem-mpsolver-impl
         Antares::optim-model-filler
